services/api/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
google-genai